Add edition_range field to work responses and refactor
This refactors the code that calculates the full range of edition years into a separate helper module so that it can be accessed by the individual `Work` endpoint as well as the Search endpoint. Also refactored is the test suite to reflect this new division of responsibilities.
