+++ lib/search.js
@@ -1,5 +1,6 @@
 const bodybuilder = require('bodybuilder')
 const { MissingParamError } = require('./errors')
+const Helpers = require('../helpers/esSourceHelpers')
 
 /*
  * Array of roles to exclude from author queries
@@ -46,7 +47,7 @@ class Search {
           if (this.reverseResult) resp.hits.hits.reverse()
           Search.formatResponsePaging(resp)
           Search.formatResponseFacets(resp)
-          Search.formatResponseEditionRange(resp)
+          Helpers.formatResponseEditionRange(resp)
           resolve(resp)
         })
         .catch(error => reject(error))
@@ -98,90 +99,6 @@ class Search {
     resp.facets = facets
     delete resp.aggregations
     /* eslint-enable no-param-reassign */
-  }
-
-  /**
-   * Generates a year range for the publication dates of the editions associated
-   * with a specific work.
-   *
-   * @param {Object} resp ElasticSearch response object
-   */
-  static formatResponseEditionRange(resp) {
-    resp.hits.hits.forEach((hit) => {
-      const startYear = this.getEditionRangeValue(hit, 'gte', 1)
-      const endYear = this.getEditionRangeValue(hit, 'lte', -1)
-
-      // eslint-disable-next-line no-param-reassign, no-underscore-dangle
-      hit._source.edition_range = `${startYear} - ${endYear}`
-    })
-  }
-
-  /**
-   * Parses the provided Work object's editions for publication dates and returns
-   * either the earliest or latest year. If not found, will return '???' for
-   * display.
-   *
-   * @param {Object} hit Work object from the ElasticSearch response
-   * @param {String} range Either 'lte' or 'gte' for the start or end of a range
-   * @param {Number} flip Either 1 or -1 to flip the sorting to ASC or DESC.
-   *
-   * @returns {String} A 4-digit representation of the year found, else ????
-   */
-  static getEditionRangeValue(hit, range, flip) {
-    let year
-    // eslint-disable-next-line no-underscore-dangle
-    const rangeInstance = hit._source.instances.sort(
-      this.startEndCompare(range, flip),
-    ).filter(instance => instance.pub_date)[0]
-    if (rangeInstance) {
-      year = new Date(`${rangeInstance.pub_date[range]}T12:00:00.000+00:00`).getFullYear()
-    } else {
-      year = '????'
-    }
-    // eslint-disable-next-line no-restricted-globals
-    if (isNaN(year)) { year = '????' }
-    return year
-  }
-
-  /**
-   * Generates a custom search function that can be customized to meet the needs
-   * of the publication date fields. Handling cases when either the full date
-   * is missing from an edition record or if one part of the date range is not
-   * present.
-   *
-   * @param {String} startEnd Either 'lte' or 'gte' for the start and end of a range
-   * @param {Number} sortFlip Either 1 or -1 to flip the sort order
-   *
-   * @returns {Function} dateComparison Function to be used by the sort method
-   */
-  static startEndCompare(startEnd, sortFlip) {
-    /**
-     * Custom implementation of a sort method to handle different cases of publication
-     * dates associated with edition records. The dates are appended with Noon GMT
-     * to ensure that no timezone issues are encountered in retrieving the year.
-     *
-     * @param {Object} a Edition object for sorting
-     * @param {Object} b Edition object for sorting
-     *
-     * @returns {Number} The sort order of the two objects, either -1, 0, or 1
-     */
-    const dateComparison = (a, b) => {
-      if (!a.pub_date && !b.pub_date) return 0
-      if (!a.pub_date) return -1 * sortFlip
-      if (!b.pub_date) return 1 * sortFlip
-
-      if (!a.pub_date[startEnd] && !b.pub_date[startEnd]) return 0
-      if (!a.pub_date[startEnd]) return -1 * sortFlip
-      if (!b.pub_date[startEnd]) return 1 * sortFlip
-
-      const d1 = new Date(`${a.pub_date[startEnd]}T12:00:00.000+00:00`).getFullYear()
-      const d2 = new Date(`${b.pub_date[startEnd]}T12:00:00.000+00:00`).getFullYear()
-
-      if (d1 > d2) return 1 * sortFlip
-      if (d1 < d2) return -1 * sortFlip
-      return 0
-    }
-    return dateComparison
   }
 
   /**
